check admin before delete

+++ src/Application/Api/Ajax/Handler/IndexAjaxHandler.php
@@ -337,7 +337,11 @@ final class IndexAjaxHandler implements AjaxHandlerInterface
                 }
                 break;
             case 'delete_play':
-                if (isset($_REQUEST['activity_id'])) {
+                if (
+                    $user instanceof User &&
+                    Access::check('interface', 100, $user->id) &&
+                    isset($_REQUEST['activity_id'])
+                ) {
                     Stats::delete((int)$_REQUEST['activity_id']);
                 }
                 ob_start();
